+++ src/tytable/_images.py
@@ -301,17 +301,6 @@ def execute_plots(
             has_header=has_header,
             data=table._data,
         )
-        if i_vals is None:
-            i_vals = resolve_i(
-                "body",
-                nhead=nhead,
-                group_positions=group_positions,
-                n_merged_body=n_merged_body,
-                has_header=has_header,
-            )
-        if i_vals is None:
-            i_vals = []
-
         body_rows = [i for i in i_vals if i > 0]
         j_vals = table._resolve_j(d.j, regex=d.regex)